Added support for MinGW

+++ include/Profiling/Timer.hpp
@@ -14,7 +14,7 @@ struct Timer
 
     float elapsed() const
     {
-        return std::chrono::duration_cast<std::chrono::nanoseconds>(std::chrono::high_resolution_clock::now() - m_Start).count() * 0.001f * 0.001f * 0.001f;
+        return static_cast<float>(std::chrono::duration_cast<std::chrono::nanoseconds>(std::chrono::high_resolution_clock::now() - m_Start).count()) * 0.001f * 0.001f * 0.001f;
     }
 
     float elapsed_millis() const
